docs: 폴더 구조 변경 및 Extension 관리
