+++ K-Keyboard/Source/Extensions/TextConstant.swift
