+++ K-Keyboard/Source/Feature/Keyboard/KeyboardView.swift
@@ -218,15 +218,3 @@ public class KeyboardView: UIView {
         sender.isSelected = !sender.isSelected
     }
 }
-
-extension UIView {
-    func addSubviews(_ views: UIView...) {
-        views.forEach { addSubview($0) }
-    }
-}
-
-extension UIStackView {
-    func addArrangedSubviews(_ views: UIView...) {
-        views.forEach { addArrangedSubview($0) }
-    }
-}
