+++ K-Keyboard/Source/Feature/Keyboard/OftenUsedCell.swift
@@ -37,21 +37,3 @@ class OftenUsedCell: UITableViewCell {
         titleLabel.leadingAnchor.constraint(equalTo: contentView.leadingAnchor, constant: 20).isActive = true
     }
 }
-
-extension UIColor {
-    private static func color(light: UIColor, dark: UIColor? = nil) -> UIColor {
-       if let dark = dark  {
-           if #available(iOS 13.0, *) {
-               return UIColor { $0.userInterfaceStyle == .dark ? dark : light }
-           }
-           return light
-       } else {
-           return light
-       }
-    }
-
-    static let systemColor: UIColor = {
-        let color = UIColor.color(light: .black, dark: .white)
-        return color
-    }()
-}
